+++ packages/insomnia/src/ui/components/modals/export-requests-modal.tsx
@@ -1,8 +1,8 @@
+import { exportRequestsToFile } from 'insomnia/src/ui/components/settings/import-export';
 import React, { type FC, type ReactNode, useEffect, useState } from 'react';
 import { Button, Checkbox, Dialog, Heading, Modal, ModalOverlay } from 'react-aria-components';
 import { useFetcher, useParams } from 'react-router';
 
-import { exportRequestsToFile } from '../../../common/export';
 import { requestGroup } from '../../../models';
 import { type GrpcRequest, isGrpcRequest } from '../../../models/grpc-request';
 import { isRequest, type Request } from '../../../models/request';
